+++ js/buildings.js
@@ -977,10 +977,10 @@ const BUILDINGS_DATA = {
             return hasInfUpgrade(9);
         },
         get autoUnlocked() {
-            return false;
+            return hasElement(283);
         },
         get noSpend() {
-            return false;
+            return hasElement(283);
         },
 
         get res() {